Ajout ordinateur & suppression des interfaces
