--- computerDatabase/src/main/java/fr/excilys/dao/DatabaseAccess.java
@@ -1,36 +1,0 @@
-package fr.excilys.dao;
-
-import java.io.FileInputStream;
-import java.io.IOException;
-import java.io.InputStream;
-import java.util.Properties;
-
-class DatabaseAccess {
-
-	protected static void getDatabaseInfo() {
-
-		final Properties properties = new Properties();
-		InputStream inputStream = null;
-
-		try {
-			inputStream = new FileInputStream("config.properties");
-
-			properties.load(inputStream);
-
-			System.out.println(properties.getProperty("db.url"));
-			System.out.println(properties.getProperty("db.username "));
-			System.out.println(properties.getProperty("db.password "));
-
-		} catch (final IOException ex) {
-			ex.printStackTrace();
-		} finally {
-			if (inputStream != null) {
-				try {
-					inputStream.close();
-				} catch (final IOException e) {
-					e.printStackTrace();
-				}
-			}
-		}
-	}
-}
