+++ computerDatabase/src/main/webapp/WEB-INF/views/addComputer.jsp
@@ -1,55 +1,63 @@
 <!DOCTYPE html>
-<html>
-<head>
-<title>Computer Database</title>
-<meta name="viewport" content="width=device-width, initial-scale=1.0">
-<!-- Bootstrap -->
-<link href="resources/css/bootstrap.min.css" rel="stylesheet" media="screen">
-<link href="resources/css/font-awesome.css" rel="stylesheet" media="screen">
-<link href="resources/css/main.css" rel="stylesheet" media="screen">
-</head>
-<body>
-    <header class="navbar navbar-inverse navbar-fixed-top">
-        <div class="container">
-            <a class="navbar-brand" href="dashboard.html"> Application - Computer Database </a>
-        </div>
-    </header>
 
-    <section id="main">
-        <div class="container">
-            <div class="row">
-                <div class="col-xs-8 col-xs-offset-2 box">
-                    <h1>Add Computer</h1>
-                    <form action="addComputer" method="POST">
-                        <fieldset>
-                            <div class="form-group">
-                                <label for="computerName">Computer name</label>
-                                <input type="text" class="form-control" id="computerName" placeholder="Computer name">
-                            </div>
-                            <div class="form-group">
-                                <label for="introduced">Introduced date</label>
-                                <input type="date" class="form-control" id="introduced" placeholder="Introduced date">
-                            </div>
-                            <div class="form-group">
-                                <label for="discontinued">Discontinued date</label>
-                                <input type="date" class="form-control" id="discontinued" placeholder="Discontinued date">
-                            </div>
-                            <div class="form-group">
-                                <label for="companyId">Company</label>
-                                <select class="form-control" id="companyId" >
-                                    <option value="0">--</option>
-                                </select>
-                            </div>                  
-                        </fieldset>
-                        <div class="actions pull-right">
-                            <input type="submit" value="Add" class="btn btn-primary">
-                            or
-                            <a href="dashboard.html" class="btn btn-default">Cancel</a>
-                        </div>
-                    </form>
-                </div>
-            </div>
-        </div>
-    </section>
-</body>
+<%@ page isELIgnored="false"%>
+<%@taglib prefix="c" uri="http://java.sun.com/jsp/jstl/core"%>
+
+<html>
+	<head>
+		<title>Computer Database</title>
+		<meta name="viewport" content="width=device-width, initial-scale=1.0">
+		<!-- Bootstrap -->
+		<link href="resources/css/bootstrap.min.css" rel="stylesheet" media="screen">
+		<link href="resources/css/font-awesome.css" rel="stylesheet" media="screen">
+		<link href="resources/css/main.css" rel="stylesheet" media="screen">
+	</head>
+	<body>
+	    <header class="navbar navbar-inverse navbar-fixed-top">
+	        <div class="container">
+	            <a class="navbar-brand" href="dashboard"> Application - Computer Database </a>
+	        </div>
+	    </header>
+
+	    <section id="main">
+	        <div class="container">
+	            <div class="row">
+	                <div class="col-xs-8 col-xs-offset-2 box">
+	                    <h1>Add Computer</h1>
+	                    <form action="addcomputer" method="POST">
+	                        <fieldset>
+	                            <div class="form-group">
+	                                <label for="computerName">Computer name</label>
+	                                <input type="text" class="form-control" name="computerName" id="computerName" placeholder="mycomputer">
+	                            </div>
+	                            <div class="form-group">
+	                                <label for="introduced">Introduced date</label>
+	                                <input type="date" class="form-control" name="introduced" id="introduced" placeholder="1970-01-01" min="1970-01-01" max="2030-01-01">
+	                            </div>
+	                            <div class="form-group">
+	                                <label for="discontinued">Discontinued date</label>
+	                                <input type="date" class="form-control" name="discontinued" id="discontinued" placeholder="1970-01-01" min="1970-01-01" max="2030-01-01">
+	                            </div>
+	                            <div class="form-group">
+	                                <label for="companyId">Company</label>
+	                                <select	class="form-control" name="companyId" id="companyId">
+	                                <option value="0" selected></option>
+										<c:forEach items="${companyList}" var="company">
+											<option value="${company.id}">
+												<c:out value="${company.name}"></c:out></option>
+										</c:forEach>
+	                                </select>
+	                            </div>                  
+	                        </fieldset>
+	                        <div class="actions pull-right">
+	                            <input type="submit" value="Add" class="btn btn-primary">
+	                            or
+	                            <a href="dashboard" class="btn btn-default">Cancel</a>
+	                        </div>
+	                    </form>
+	                </div>
+	            </div>
+	        </div>
+	    </section>
+	</body>
 </html>
